+++ src/sccz80/expr.h
@@ -1,6 +1,6 @@
 /* expr.c */
 extern void ClearCast(LVALUE *lval);
-extern int expression(int *con, int *val);
+extern int expression(int *con, double *val);
 extern int heir1(LVALUE *lval);
 extern int heir1a(LVALUE *lval);
 extern int heir2a(LVALUE *lval);
